switch to nextjs pages router
