--- web/src/app/not-found.css
@@ -1,3 +1,0 @@
-.not-found {
-  font-size: 1.2rem;
-}
