+++ web/src/components/ui/Button.module.css
@@ -36,7 +36,7 @@
   }
 }
 
-button:disabled {
+.button:disabled {
   opacity: 0.8;
   background-color: grey;
   border: none;
