+++ web/src/pages/_app.tsx
@@ -1,0 +1,19 @@
+import "./globals.css";
+import type { AppProps } from "next/app";
+import { AuthProvider } from "@/hooks/useAuth";
+import Head from "next/head";
+import Layout from "@/components/layout/Layout";
+
+export default function App({ Component, pageProps }: AppProps) {
+  return (
+    <AuthProvider>
+      <Head>
+        <title>ARC Services Portal | UCL</title>
+        <meta property="description" content="ARC Services Portal" key="description" />
+      </Head>
+      <Layout>
+        <Component {...pageProps} />
+      </Layout>
+    </AuthProvider>
+  );
+}
