+++ web/src/pages/globals.css
