+++ web/src/pages/profile/approved-researcher/components/AgreementText.module.css
