+++ web/src/pages/profile/approved-researcher/components/AgreementText.tsx
@@ -1,6 +1,4 @@
-"use client";
-
-import "./AgreementText.css";
+import "./AgreementText.module.css";
 import Markdown from "react-markdown";
 
 type AgreementTextProps = {
