+++ web/src/pages/profile/components/Profile.tsx
@@ -1,6 +1,4 @@
-"use client";
-
-import { useAuth } from "../../hooks/useAuth";
+import { useAuth } from "@/hooks/useAuth";
 import LoginFallback from "@/components/ui/LoginFallback";
 import Button from "@/components/ui/Button";
