+++ web/src/pages/studies/index.tsx
